move accounts task action to actions/ directory
